Fix some Sonar issues.
